[vm] Basic bytecode loading. `Myst::VM::BytecodeBuffer` will be the program storage for the VM's interpreter. It can be initialized empty, from a file, or from an arbitrary IO, and can also be appended by other buffers to create a single buffer for an entire program.
A BytecodeBuffer is simply a wrapper around an array of `Instruction`s, which are the individual commands of a bytecode. Each Instruction has an `InstructionType`, specifying the command it represents, and a variable number of arguments, depending on the semantics of the type of the instruction.

For now, all arguments are kept as strings. Eventually, literal arguments and other types will need to be parsed correctly into Value types, identifiers, etc.