[interpreter] Better scopes with chains, argument order for functions. Scopes are now managed as a reverse linked list, rather than an array, where each scope has a reference to a parent, allowing it to maintain proper lookup semantics when classes and modules are implemented.
Scopes can also be marked as "restrictive", which maens that it acts as the "top-level scope" for all lookups. Symbols not found in that scope will be considered undefined.

Arguments to functions were also being inserted in reverse order (`func(a, b)` was treated as `func(b, a)`), which is now fixed as well.