[parser,interpreter] Key interpolation for Maps. Maps now support key-interpolation: expressions that are evaluated to determine the key of a Map. Key interpolation is done using a postfix_expression wrapped in angle brackets (`<postfix_expression>`). As such, any access expressions, function calls, or literals can be used directly in interpolation. To use complex expressions like binary operators, wrap the expression in parenthesis as well (`<(binary_expression)>`).
The advantage of this syntax is that maps only have one key-value form - the suffix colon. This is opposed to languages like Ruby, where Hashes can use either the suffix-colon _or_ the hashrocket as the key-value form. The usage of the angle bracket syntax, when written idiomatically, is also the same number of characters as the idiomatic hashrocket, but I think it reads better in general.