Update fibonnaci example to use guard clauses.
Using `return if some_condition` at the start of a method should be considered an anti-pattern if the same expression can be written as a guard clause.

Not only would this create a more semantic error (`NoFunctionMatch`, vs. `no method for nil` somewhere down the road), but it's also more concise.