[interpreter] Member access and working modules. `MemberAccessExpressions` are now properly supported. The member specifier must be an identifier for now (a possible workaround would be supporting the JS-like array access method as well `receiver[member]`, but I'm not particularly keen on multiple ways to do the same simple thing).
This support also means that Modules can be used properly. `examples/modules.mt` now contains a working example of an `IO` module with another module nested inside of it. It demonstrates scoping rules. For now, nested module call syntax is a simple `MemberAccessExpression`. I'm torn between whether to keep this style (similar to Elixir), or allow the C-like `::` operator (similar to Ruby). I _think_ I prefer the dot syntax; even though the syntax highlighting I use does not support it, I find it legible and more consistent. This is especially true for constants, where in Ruby they are normally accessed using the namespace operator (i.e., `Nested::Modules.CONSTANT`), while it does convey a specific meaning, it's not immediately obvious that this is the same as `Nested::Module.CONSTANT`, which is also supported. I'd rather just have one method that is always consistent.