[kernel,interpreter] Add Unary expression evaluation, `IO#gets`, improve native functor definitions. Native functors can now either be defined inline in `native_func`, or by referencing an existing method to use instead. The native methods on `IO` utilize this to be able to access other private methods defined in that module.
This comes from the addition of `IO#gets` in the standard library. The implementation is not great and will result in large numbers of system calls (it reads one character at a time :/). Much of this should be avoidable by utilizing more of Crystal's native `read_line`, `read_fully`, and other methods.