[lexer,parser,interpreter] Add support for `include` in module definitions. `include` inserts the given module at the beginning of the "ancestor chain" of the current scope. That means any values defined in the _included_ module may be looked up from the _including_ module without a namespaced reference.
See `examples/include.mt` for more information and a detailed look at the behavior.

Adding this support required a fix to how lookups are performed. Previously, lookup only checked the stack of scopes that were directly in the symbol table. This often (but not always...) ignored parent scopes and would cause lookup errors when requesting members out-of-line (e.g., via a variable reference to a module).