[lexer,parser,interpreter] Introduce `when` chains as a replacement for `if/elif`. `when` is a surprisingly good substitute for both `if` and `elif`. The semantics of `when` are essentially the same as they are in Ruby, but without the required `case` beforehand.
The rationale for ditching `if` entirely is an attempt at limiting keywords and consistency. There is no reason to have multiple constructs that essentially do the same thing, where the only difference is that `if` can only be followed by an `else if` construct for chaining, which isn't always clear, especially to new programmers.