[parser] Refactor AST nodes into distinct classes. All AST node types are now all defined manually (no more macros). This should be slightly faster to compile, and also allows much greater flexibility with how the types and their properties are defined and initialized.
With this refactor, all nodes now also have ample documentation as to their use and the syntax used to denote them. The hope here is that anyone reading through the AST can fully understand both the _how_ and the _why_ for every node, even with no other knowledge of the language.