[lexer,parser] Support for parsing List literals The parser now understands the structure of list literals and properly generates ListLiteral nodes.
This addition uncovered a few subtle behaviors as well. Notably, `Lexer#finished?` returns true as soon as EOF is reached (not consumed), which caused `Lexer#lex_all` to terminate early and potentially miss tokens at the end of a stream.

Also, this commit changes the contract for parsing methods in `Parser`. A given method should only consume the exact tokens necessary to formulate the node it generates. This should avoid some of the newline and spacing issues that came up in the original parser, at the expense of somewhat-less-elegant/abstract code at higher levels.