[parser,spec] Implement When/Unless chains. Something I noticed while writing the specs for When expressions is that they are ambiguous for nesting. That is, writing a `when` within a `when` will always be parsed as a `when` _followed by_ a `when`. The nesting structure will always be lost.
This is inherent to the syntax, and cannot be addressed in a single-pass parse. I don't think this is necessarily a problem, however, as nesting conditionals in that way is normally indicative of an anti-pattern. Additionally, `when`s that are scoped within some other construct are _not_ ambiguous and will parse just fine.