[interpreter] Refactor TFunctor to contain both natural and native clauses. TFunctor and TNativeFunctor have become TFunctorDef and TNativeDef, both of which inherit from the new Callable intermediary. TFunctorDef is a simple wrapper around Def, while TNativeDef is a wrapper around a Proc structure.
TFunctor is now a generic container of Callables, also tracking a lexical scope and potential parent functor. Functor parenting is essentially a mini version of vtables in C++, but allowing arbitrary clauses and redefinitions. This will probably be removed at some point in favor of a simpler approach that just does recursive lookup through the scope stack.

Argument matching with TNativeDefs still needs some more thought. Right now, the definition simply accepts an Array of arguments. This is fine, but makes adding new clauses with different argument counts and such very difficult/impossible. A way of defining the expectations of the definition _before_ calling into it will make this much easier.