[parser,interpreter,spec] Add support for Splats in Patterns. Patterns for matches can now include Splats to collect remaining values after the rest of the pattern has been matched.
Adding this support also revealed an issue in the Parser and AST where Splats were not being compared properly (spurious `def_equals_and_hash` with no arguments) and were not `case`d in the `to_pattern` utility in the parser. Both of these issues have been addressed and are now properly tested.