[interpreter,spec] Add `initialize` callback, properly support IVars. Instantiations now call `initialize` on the newly-created instance, so long as a functor with that name exists for it. The initializer is given the arguments and block (without modification) from the Instantiation.
Additionally, IVars now properly operate on `current_self` rather than `current_scope`, and are stored in a distinct property of a value (outside of the `scope` of the value, called `ivars`). Attempting to read an IVar that has not yet been initialized will initialize it with a value of `nil`.