[lexer,parser,spec] Start support for exception handling. The parser now understands the new `raise` keyword and that it must be given an argument to be valid. Additionally, the `Raise`, `Rescue`, and `ExceptionHandler` nodes have been added.
The `ExceptionHandler` node is a wrapper for Expressions that are capable of handling exceptions (e.g., a Def or Block). Whenever a `rescue` is encountered at the end of a capable Expressions, it is wrapped in an ExceptionHandler and the rescue/else/ensure blocks will then be parsed and added accordingly.