[interpreter] Add "Kernel" as the parent scope for native types. This was the original reason for #18. Take the following code:
```
defmodule Enumerable
end

deftype List
  include Enumerable
end
```

This used to fail because the `List` type did not have a parent scope, and so it could not find the `Enumerable` module to include. Now, with the dynamic types for values, the Kernel scope can be injected as the parent for each instance of the Interpreter. That allows the above code to properly resolve `Enumerable` and work as expected.