[lexer,parser] (#22) Contextual lexing and string interpolations. The lexer now tracks two additional values while operating: the brace stack and the context stack.
The brace stack is used to syntactically match opening and closing braces, and to assert that all higher braces are matched when closing a brace. For example, the code `{[}` is syntactically invalid, because the closing curly brace appears before the opened square brace has been matched with a closing square brace.

The context stack is the biggest change to the lexer in quite a while. Based on the current context (the top of the stack), the lexer can analyze the source differently and create different tokens from the stream. This is used (and somewhat required) to allow for string interpolations. Inside of a string, the sequence `<(` should be lexed as the start of an interpolation, inside of which parsing operates as it does _outside_ of a string, but string lexing is resumed as soon as the closing `)>` is matched. Again, this uses the brace stack to ensure syntactic validity. The context _stack_ allows arbitrary nesting of strings, interpolations, and normal source material.

Lastly, InterpolatedStringLiterals are now part of the AST, and are properly generated based on a token stream containing INTERP_START and INTER_END tokens. The final step for supporting string interpolations will be visiting these nodes in the interpreter.