[parser] (#22) Only allow single expressions inside of `InterpolatedStringLiterals`. Interpolations may contain arbitrary expressions, but only one (though the expression may be a Def or Block, which allows a body of multiple expressions).
Additionally, the parser can now distinguish an interpolation containing only a string literal, and a string literal piece of an InterpolatedStringLiteral. For example, the interpolated string: `"<("hi")>"` will keep the interpolation component, even though the value being interpolated is itself a string literal. This is to preserve the semantics where interpolations always call `.to_s` on the result of the interpolated expression.