[nativelib,stdlib] Add `File.write(data)`. `File.write` is similar to `IO.print`, but writing to a file instead of to the terminal.
`File.write` relies on `FSUtils.write`, which does the actual writing. Writing a file depends on the file being opened in a writable mode. To accomodate this while maintaining read-only mode as well, `FSUtils.open` now expects a `mode` argument, and `File.open` now accepts an optional `mode` argument that will default to `"r"` for read-only. Using `"w"` or `"a"` will open the file in read mode.