[interpreter,spec] (#109) Fix `self` and scope resolution for blocks on static methods. `Invocation` was incorrectly determining the value of `self` to use when calling a function. Before #131, it was only checking if the Call had a receiver, and would use that value if it did. After #131, if the Call did _not_ have a receiver, it would then check if the functor was a closure, and use the closed value of `self` if it was.
Now, Invocations _always_ perform both checks, with closures taking higher precedence than Call receivers (though, both will be pushed to the stack if the conditions are met).

Before this commit, I was very confused why static methods weren't working while non-static methods were. I assumed it was an issue with `scope` vs `instance_scope` on Types or something like that, but had no idea.

Now, though, I'm even more confused how _any_ case was working previously, as the Call receiver was _always_ being pushed as the value of `self`. At least with this version of the code, the value of `self` that will be present is more explicit and well-specified. ¯\_(ツ)_/¯