[semantic] Refactor SemanticVisitor into Semantic module, add ParamNamesAssertion. Everything relating to semantic analysis now lives under the `Semantic` module, this should avoid name conflicts with the interpreter, as well as provide a more obvious structure in general.
This commit also introduces the idea of an `Assertion` for the semantic visitor. Rather than having assertion about semantics be made directly as part of the visitor, assertions will now be made as separate classes that the visitor can call out to at any point. This should enable better code re-use, as well as follow the Single Responsibility principle much better.