[parser] Implement doc parsing for block-level expressions. AST nodes now have a `doc` property that contains any documentation for that node.
The documentation for a node is the concatenation of every comment line that directly precedes the node. A blank line of padding is the end of the documentation block from the bottom up. For example:

```
def foo; end

def foo; end

def foo; end

def foo; end
```

Newlines within the doc comment are preserved.