[lexer,parser,spec] Make doc comments an explicit language construct. Doc comments are now a special language construct beyond normal comments. Instead of trying to guess at which comments count as documentation, Myst now uses a "special comment" structure to explicitly indicate documentation.
To create a doc comment, use `#doc` followed by a reference to the method/type/module/property being documented, and optionally a "return value" of the method. Then, write the content of the documentation on the following lines by starting them with `#|`. All together, a full doc comment could look like this:

```
def each(&block)
  # ...
end
```

This special structure provides a few main benefits:
- Documentation within a codebase is more obvious and called out.
- Highlighters and code analyzers can expect a consistent structure for documentation.
- Documentation can be written anywhere in the code base and attached to objects defined anywhere else.

This commit only adds the lexing and parsing of doc comments and does not yet attach them to runtime values.