[interpreter,spec] Collect documentation during program execution. The interpreter now knows how to interpret DocComments, and stores their content in a flat table (the `doc_table`) in the interpreter itself. Users can access documentation for any module, type, method, or property at run time using the `doc` method and providing the absolute path to the value they want. For example, getting documentation on the assert library's `is_falsey` would look like `Assert.Assertion#is_falsey`.
I'd like to improve this to allow documentation lookup based on a value. That would help users who are _not_ familiar with a library learn more about it through a REPL or other runtimes. Since that's the primary audience for documentation, it makes sense to try to support their use case the most.