[docs] Start documenting native methods in special `stdlib/native` folder. The native folder acts like a "magic" directory that contains no implementation code, but provides objects for the documentation generator to scan for and parse into the documentation blob for the standard library.
Using this approach allows the generator to create a single structure for the entire library exposed by Myst, without regard for the implementations of the methods themselves.

Files in the `stdlib/native` folder should never be included by Myst programs, as they will be redundant and may possibly cause errors depending on the content of the file.