[parser] Support unions in type restrictions Type restrictions (both for parameters and return types) can now specify a union of two or more types as the restriction.
This should allow users to condense multiple clauses with equivalent functionality into a single clause accepting multiple types for a parameter/return type, without loosening the accuracy of the type requirements.

One note in this change is that unions restrictions for block parameters are absolutely ambiguous. There is no non-fundamental change that would be able to resolve the ambiguity between the pipe character used for unioning two types and the pipe character used to end the block parameter list.

To address this, union restrictions in block parameters are only allowed if the restriction is wrapped in parentheses. This is not a final decision, but simply one that punts the decision down the road on what to do with unions in general.

Unions are not yet understood by the interpreter.